Allow to set absolute volume
This allows to set the absolute volume of spotify. Note that sometimes you'll
run into a off-by-one error as spotify returns a volume that is often one higher
than the one you actually set. For example:

```javascript
spotify.setVolume(42, function() {
  spotify.getState(function(err, state) {
    console.log(state.volume); // will output: 41
  });
});
```

This shouldn't have any noticable effect for the user of the library though.