Implement jumpTo method
This implements a jumpTo method to jump to a specific position of the song.
Example:

```javascript
spotify.jumpTo(15, function() {
  console.log('Jumped 15th second of the song');
});
```

The test is not optimal now as we need a timeout to give spotify some time
to catch up from the beforeEach "playTrack", but I wanted to keep this change
small.