Escape quotes when getting track information
If you have a song title that includes quotes the old version produced
a parse error. For example for "spotify:track:4EZz8Byhbjk0tOKFJlCgPB" you'll run
into this, because it includes 7" in the album title.

With this commit all the quotes in the values getting properly escaped.

This also extracts the get_track script to a seperate file as it makes
maintaining easier.